add fortnite config file

+++ src/com/tccsafeo/entities/PlayerData.java
@@ -2,13 +2,13 @@ package com.tccsafeo.entities;
 
 public class PlayerData {
     public String key;
-    public Integer value;
+    public Double value;
 
     public PlayerData() {
         super();
     }
 
-    public PlayerData(String key, Integer value) {
+    public PlayerData(String key, Double value) {
         this.key = key;
         this.value = value;
     }
